third_party/WebKit/LayoutTests/fast/dom/shadow/host-context-class-change-expected.txt

Class change affecting a node in a distributed node's subtree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(b, null).backgroundColor is transparent
PASS getComputedStyle(b, null).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
